requirements.txt
streamlit
google-genai
PyPDF2
pandas
altair